Start of messaging
Start creating the messaging, Nokia section now displaying json containing the user access token and secret key. Still need to figure out how to get that to the ios application - one option is to implement firebase into the go lang application but that should be a last results since we can try to get current working.